+++ src/util.rs
@@ -252,7 +252,17 @@ pub fn get_theme(opt: &Opt, config: &Config) -> ConfigTheme {
     match theme {
         ConfigTheme::Auto => {
             if atty::is(Stream::Stdout) && atty::is(Stream::Stderr) && atty::is(Stream::Stdin) {
-                let timeout = Duration::from_millis(100);
+                let minimum_timeout = Duration::from_millis(100);
+                let timeout = if let Ok(latency) = termbg::latency(Duration::from_millis(10000)) {
+                    if latency * 2 > minimum_timeout {
+                        latency * 2
+                    } else {
+                        minimum_timeout
+                    }
+                } else {
+                    minimum_timeout
+                };
+
                 if let Ok(theme) = termbg::theme(timeout) {
                     match theme {
                         termbg::Theme::Dark => ConfigTheme::Dark,